Connect to a League

This function creates a connection object which stores parameters and gets a login-cookie if available - it does so by passing arguments to the appropriate league-based handler.

ff_connect(platform = "mfl", league_id, ...)

Arguments

platform
one of MFL or Sleeper (Fleaflicker, ESPN, Yahoo in approximate priority order going forward)

league_id
league_id (currently assuming one league at a time)

...
other parameters passed to the connect function for each specific platform.

rate_limit Defaults to TRUE. Pass FALSE to turn off.

rate_limit_number number of attempts to try in rate_limit_seconds.

APIKEY MFL-specific - grants access to perform something as a user in a specific league

user_name Collects user_name field for leagues that support it (MFL, Sleeper so far)

password Collects password and attempts to fetch an authorization token which functions a lot like an APIKEY

user_agent Identifies user scraping the data

Value

a list that stores MFL connection objects