Get Draft Picks

Returns all draft picks (current and future) that belong to a specific franchise and have not yet been converted into players (i.e. selected.)

Usage

ff_draftpicks(conn, ...)

# S3 method for class 'espn_conn'
ff_draftpicks(conn, ...)

# S3 method for class 'flea_conn'
ff_draftpicks(conn, franchise_id = NULL, ...)

# S3 method for class 'mfl_conn'
ff_draftpicks(conn, ...)

# S3 method for class 'sleeper_conn'
ff_draftpicks(conn, ...)

Arguments

conn

the list object created by ff_connect()

...

other arguments (currently unused)

franchise_id

A list of franchise IDs to pull, if NULL will return all franchise IDs

Value

Returns a dataframe with current and future draft picks for each franchise

Methods (by class)

ff_draftpicks(espn_conn): ESPN: does not support future/draft pick trades - for draft results, please use ff_draft.

ff_draftpicks(flea_conn): Fleaflicker: retrieves current and future draft picks, potentially for a specified team.

ff_draftpicks(mfl_conn): MFL: returns current and future picks

ff_draftpicks(sleeper_conn): Sleeper: retrieves current and future draft picks

Examples

# \donttest{
try({ # try only shown here because sometimes CRAN checks are weird
  conn <- espn_connect(
    season = 2018,
    league_id = 1178049,
    espn_s2 = Sys.getenv("TAN_ESPN_S2"),
    swid = Sys.getenv("TAN_SWID")
  )

  ff_draftpicks(conn)
}) # end try
#> Warning: ESPN does not support draft pick trades. For draft results, please use ff_draft()
#> NULL
# }

# \donttest{
try({ # try only shown here because sometimes CRAN checks are weird
  conn <- fleaflicker_connect(2020, 206154)
  ff_draftpicks(conn, franchise_id = 1373475)
}) # end try
# }

# \donttest{
try({ # try only shown here because sometimes CRAN checks are weird
  dlf_conn <- mfl_connect(2020, league_id = 37920)
  ff_draftpicks(conn = dlf_conn)
}) # end try
#> # A tibble: 80 × 5
#>    season franchise_id franchise_name round original_franchise_id
#>     <dbl> <chr>        <chr>          <dbl> <chr>                
#>  1   2021 0001         @dibari22          1 0001                 
#>  2   2021 0001         @dibari22          2 0001                 
#>  3   2021 0001         @dibari22          3 0001                 
#>  4   2021 0001         @dibari22          5 0001                 
#>  5   2021 0001         @dibari22          2 0009                 
#>  6   2021 0002         OPEN               1 0002                 
#>  7   2021 0002         OPEN               2 0002                 
#>  8   2021 0002         OPEN               3 0002                 
#>  9   2021 0002         OPEN               4 0002                 
#> 10   2021 0002         OPEN               5 0002                 
#> # ℹ 70 more rows
# }

# \donttest{
try({ # try only shown here because sometimes CRAN checks are weird
  jml_conn <- ff_connect(platform = "sleeper", league_id = "522458773317046272", season = 2020)
  ff_draftpicks(jml_conn)
}) # end try
#> # A tibble: 144 × 5
#>    season franchise_id franchise_name round original_franchise_id
#>    <chr>         <int> <chr>          <int>                 <int>
#>  1 2021              1 Fake News          3                     1
#>  2 2021              1 Fake News          4                     1
#>  3 2022              1 Fake News          2                     1
#>  4 2022              1 Fake News          3                     1
#>  5 2022              1 Fake News          4                     1
#>  6 2023              1 Fake News          1                     1
#>  7 2023              1 Fake News          2                     1
#>  8 2023              1 Fake News          3                     1
#>  9 2023              1 Fake News          4                     1
#> 10 2021              2 KingGabe           1                     2
#> # ℹ 134 more rows
# }