Get League Rosters

This function returns a tidy dataframe of team rosters

ff_rosters(conn, ...)

# S3 method for flea_conn
ff_rosters(conn, ...)

# S3 method for mfl_conn
ff_rosters(conn, custom_players = FALSE, ...)

# S3 method for sleeper_conn
ff_rosters(conn, ...)

Arguments

conn
a conn object created by ff_connect()

...
arguments passed to other methods (currently none)

custom_players
TRUE or FALSE - include custom players? defaults to FALSE

Value

A tidy dataframe of rosters, joined to basic player information and basic franchise information

Methods (by class)

flea_conn: Fleaflicker: Returns roster data (minus age as of right now)

mfl_conn: MFL: returns roster data

sleeper_conn: Sleeper: Returns all roster data.

Examples

# \donttest{
joe_conn <- ff_connect(platform = "fleaflicker", league_id = 312861, season = 2020)
ff_rosters(joe_conn)

#> # A tibble: 357 x 7
#>    franchise_id franchise_name player_id player_name  pos   team  sportradar_id 
#>           <int> <chr>              <int> <chr>        <chr> <chr> <chr>         
#>  1      1578553 Running Bear       12032 Carson Wentz QB    PHI   e9a5c16b-4472…
#>  2      1578553 Running Bear        7378 Cam Newton   QB    NE    214e55e4-a089…
#>  3      1578553 Running Bear       15622 Joshua Kell… RB    LAC   62542e04-3c44…
#>  4      1578553 Running Bear       13358 Matt Breida  RB    MIA   6249d2c0-75dc…
#>  5      1578553 Running Bear        7369 A.J. Green   WR    CIN   c9701373-23f6…
#>  6      1578553 Running Bear       13782 Anthony Mil… WR    CHI   bfaedf99-7618…
#>  7      1578553 Running Bear        9338 Robert Woods WR    LAR   618bedee-9259…
#>  8      1578553 Running Bear       13209 Tim Patrick  WR    DEN   2a443351-5e63…
#>  9      1578553 Running Bear       13874 Ian Thomas   TE    CAR   ed8a8fd2-df67…
#> 10      1578553 Running Bear       13489 Robert Tony… TE    GB    7c9c7800-69d0…
#> # … with 347 more rows

# }
# \donttest{
ssb_conn <- ff_connect(platform = "mfl", league_id = 54040, season = 2020)
ff_rosters(ssb_conn)

#> # A tibble: 443 x 11
#>    franchise_id franchise_name player_id player_name pos   team    age
#>    <chr>        <chr>          <chr>     <chr>       <chr> <chr> <dbl>
#>  1 0001         Team Pikachu   13189     Engram, Ev… TE    NYG    26.2
#>  2 0001         Team Pikachu   11680     Landry, Ja… WR    CLE    28  
#>  3 0001         Team Pikachu   13645     Smith, Tre… WR    NOS    24.9
#>  4 0001         Team Pikachu   12110     Brate, Cam… TE    TBB    29.4
#>  5 0001         Team Pikachu   13168     Reynolds, … WR    LAR    25.8
#>  6 0001         Team Pikachu   13793     Valdes-Sca… WR    GBP    26.1
#>  7 0001         Team Pikachu   13879     Allen, Kyle QB    WAS    24.7
#>  8 0001         Team Pikachu   13377     Patrick, T… WR    DEN    27  
#>  9 0001         Team Pikachu   13254     Shaheen, A… TE    MIA    26.9
#> 10 0001         Team Pikachu   8673      Ginn Jr., … WR    FA     35.6
#> # … with 433 more rows, and 4 more variables: roster_status <chr>,
#> #   drafted <chr>, draft_year <chr>, draft_round <chr>

# }

# \donttest{
jml_conn <- ff_connect(platform = "sleeper", league_id = "522458773317046272", season = 2020)
ff_rosters(jml_conn)

#> # A tibble: 350 x 7
#>    franchise_id franchise_name player_id player_name     pos   team    age
#>    <chr>        <chr>          <chr>     <chr>           <chr> <chr> <dbl>
#>  1 1            Fake News      1110      T.Y. Hilton     WR    IND    31  
#>  2 1            Fake News      1339      Zach Ertz       TE    PHI    30  
#>  3 1            Fake News      1426      DeAndre Hopkins WR    ARI    28.5
#>  4 1            Fake News      1825      Jarvis Landry   WR    CLE    28  
#>  5 1            Fake News      2025      Albert Wilson   WR    MIA    28.4
#>  6 1            Fake News      2197      Brandin Cooks   WR    HOU    27.2
#>  7 1            Fake News      2319      DeVante Parker  WR    MIA    27.8
#>  8 1            Fake News      232       Frank Gore      RB    NYJ    37.5
#>  9 1            Fake News      2822      Adam Humphries  WR    TEN    27.4
#> 10 1            Fake News      289       Drew Brees      QB    NO     41.9
#> # … with 340 more rows

# }