Get League-Specific Scoring History

(Experimental!) This function reads your league's ff_scoring rules and maps them to nflfastr week-level data. Not all of the scoring rules from your league may have nflfastr equivalents, but most of the common ones are available!

Usage

ff_scoringhistory(conn, season, ...)

# S3 method for espn_conn
ff_scoringhistory(conn, season = 1999:nflreadr::most_recent_season(), ...)

# S3 method for flea_conn
ff_scoringhistory(conn, season = 1999:nflreadr::most_recent_season(), ...)

# S3 method for mfl_conn
ff_scoringhistory(conn, season = 1999:nflreadr::most_recent_season(), ...)

# S3 method for sleeper_conn
ff_scoringhistory(conn, season = 1999:nflreadr::most_recent_season(), ...)

# S3 method for template_conn
ff_scoringhistory(conn, season = 1999:nflreadr::most_recent_season(), ...)

Arguments

conn

a conn object created by ff_connect()

season

season a numeric vector of seasons (earliest available year is 1999)

...

other arguments

Value

A tidy dataframe of weekly fantasy scoring data, one row per player per week

Methods (by class)

ff_scoringhistory(espn_conn): ESPN: returns scoring history in a flat table, one row per player per week.

ff_scoringhistory(flea_conn): Fleaflicker: returns scoring history in a flat table, one row per player per week.

ff_scoringhistory(mfl_conn): MFL: returns scoring history in a flat table, one row per player per week.

ff_scoringhistory(sleeper_conn): Sleeper: returns scoring history in a flat table, one row per player per week.

ff_scoringhistory(template_conn): template: returns scoring history in a flat table, one row per player per week.

See also

Examples

# \donttest{
try({ # try only shown here because sometimes CRAN checks are weird
  conn <- espn_connect(season = 2020, league_id = 899513)
  ff_scoringhistory(conn, season = 2020)
}) # end try
#> # A tibble: 5,397 × 24
#>    season  week gsis_id    sportradar_id  espn_id player_name pos   team  points
#>     <int> <int> <chr>      <chr>          <chr>   <chr>       <chr> <chr>  <dbl>
#>  1   2020    15 00-0034490 83849bc5-0b6c… 3894901 Ezekiel Tu… LB    ARI      3.1
#>  2   2020     1 00-0035752 23461354-f2d7… 3040206 Chris Stre… QB    ARI      0.3
#>  3   2020    17 00-0035752 23461354-f2d7… 3040206 Chris Stre… QB    ARI      7.4
#>  4   2020     1 00-0035228 dd5a6b6e-ffae… 3917315 Kyler Murr… QB    ARI     26.3
#>  5   2020     2 00-0035228 dd5a6b6e-ffae… 3917315 Kyler Murr… QB    ARI     32.1
#>  6   2020     3 00-0035228 dd5a6b6e-ffae… 3917315 Kyler Murr… QB    ARI     21.7
#>  7   2020     4 00-0035228 dd5a6b6e-ffae… 3917315 Kyler Murr… QB    ARI     23.1
#>  8   2020     5 00-0035228 dd5a6b6e-ffae… 3917315 Kyler Murr… QB    ARI     26.3
#>  9   2020     6 00-0035228 dd5a6b6e-ffae… 3917315 Kyler Murr… QB    ARI     28.9
#> 10   2020     7 00-0035228 dd5a6b6e-ffae… 3917315 Kyler Murr… QB    ARI     37.1
#> # ℹ 5,387 more rows
#> # ℹ 15 more variables: interceptions <dbl>, passing_2pt_conversions <dbl>,
#> #   passing_tds <dbl>, passing_yards <dbl>, receiving_2pt_conversions <dbl>,
#> #   receiving_fumbles_lost <dbl>, receiving_tds <dbl>, receiving_yards <dbl>,
#> #   receptions <dbl>, rushing_2pt_conversions <dbl>,
#> #   rushing_fumbles_lost <dbl>, rushing_tds <dbl>, rushing_yards <dbl>,
#> #   sack_fumbles_lost <dbl>, special_teams_tds <dbl>
# }

# \donttest{
try({ # try only shown here because sometimes CRAN checks are weird
  conn <- fleaflicker_connect(2020, 312861)
  ff_scoringhistory(conn, season = 2020)
}) # end try
#> # A tibble: 5,392 × 24
#>    season  week gsis_id    sportradar_id  fleaflicker_id player_name pos   team 
#>     <int> <int> <chr>      <chr>          <chr>          <chr>       <chr> <chr>
#>  1   2020     1 00-0035752 23461354-f2d7… NA             Chris Stre… QB    ARI  
#>  2   2020    17 00-0035752 23461354-f2d7… NA             Chris Stre… QB    ARI  
#>  3   2020     1 00-0035228 dd5a6b6e-ffae… NA             Kyler Murr… QB    ARI  
#>  4   2020     2 00-0035228 dd5a6b6e-ffae… NA             Kyler Murr… QB    ARI  
#>  5   2020     3 00-0035228 dd5a6b6e-ffae… NA             Kyler Murr… QB    ARI  
#>  6   2020     4 00-0035228 dd5a6b6e-ffae… NA             Kyler Murr… QB    ARI  
#>  7   2020     5 00-0035228 dd5a6b6e-ffae… NA             Kyler Murr… QB    ARI  
#>  8   2020     6 00-0035228 dd5a6b6e-ffae… NA             Kyler Murr… QB    ARI  
#>  9   2020     7 00-0035228 dd5a6b6e-ffae… NA             Kyler Murr… QB    ARI  
#> 10   2020     9 00-0035228 dd5a6b6e-ffae… NA             Kyler Murr… QB    ARI  
#> # ℹ 5,382 more rows
#> # ℹ 16 more variables: points <dbl>, interceptions <dbl>,
#> #   passing_2pt_conversions <dbl>, passing_first_downs <dbl>,
#> #   passing_tds <dbl>, passing_yards <dbl>, receiving_2pt_conversions <dbl>,
#> #   receiving_first_downs <dbl>, receiving_tds <dbl>, receiving_yards <dbl>,
#> #   receptions <dbl>, rushing_2pt_conversions <dbl>, rushing_first_downs <dbl>,
#> #   rushing_tds <dbl>, rushing_yards <dbl>, special_teams_tds <dbl>
# }

# \donttest{
try({ # try only shown here because sometimes CRAN checks are weird
  ssb_conn <- ff_connect(platform = "mfl", league_id = 54040, season = 2020)
  ff_scoringhistory(ssb_conn, season = 2020)
}) # end try
#> # A tibble: 5,392 × 26
#>    season  week gsis_id    sportradar_id   mfl_id player_name pos   team  points
#>     <int> <int> <chr>      <chr>           <chr>  <chr>       <chr> <chr>  <dbl>
#>  1   2020     1 00-0035752 23461354-f2d7-… 15060  Chris Stre… QB    ARI      0.8
#>  2   2020    17 00-0035752 23461354-f2d7-… 15060  Chris Stre… QB    ARI      7.9
#>  3   2020     1 00-0035228 dd5a6b6e-ffae-… 14056  Kyler Murr… QB    ARI     28.3
#>  4   2020     2 00-0035228 dd5a6b6e-ffae-… 14056  Kyler Murr… QB    ARI     33.6
#>  5   2020     3 00-0035228 dd5a6b6e-ffae-… 14056  Kyler Murr… QB    ARI     21.2
#>  6   2020     4 00-0035228 dd5a6b6e-ffae-… 14056  Kyler Murr… QB    ARI     30.6
#>  7   2020     5 00-0035228 dd5a6b6e-ffae-… 14056  Kyler Murr… QB    ARI     28.3
#>  8   2020     6 00-0035228 dd5a6b6e-ffae-… 14056  Kyler Murr… QB    ARI     35.9
#>  9   2020     7 00-0035228 dd5a6b6e-ffae-… 14056  Kyler Murr… QB    ARI     43.6
#> 10   2020     9 00-0035228 dd5a6b6e-ffae-… 14056  Kyler Murr… QB    ARI     46.9
#> # ℹ 5,382 more rows
#> # ℹ 17 more variables: interceptions <dbl>, passing_2pt_conversions <dbl>,
#> #   passing_tds <dbl>, passing_yards <dbl>, receiving_2pt_conversions <dbl>,
#> #   receiving_fumbles_lost <dbl>, receiving_tds <dbl>, receiving_yards <dbl>,
#> #   receptions <dbl>, rushing_2pt_conversions <dbl>, rushing_first_downs <dbl>,
#> #   rushing_fumbles_lost <dbl>, rushing_tds <dbl>, rushing_yards <dbl>,
#> #   sack_fumbles_lost <dbl>, special_teams_tds <dbl>, …
# }

# \donttest{
try({ # try only shown here because sometimes CRAN checks are weird
  conn <- ff_connect(platform = "sleeper", league_id = "522458773317046272", season = 2020)
  ff_scoringhistory(conn, season = 2020)
}) # end try
#> # A tibble: 5,417 × 23
#>    season  week gsis_id  sportradar_id sleeper_id player_name pos   team  points
#>     <int> <int> <chr>    <chr>         <chr>      <chr>       <chr> <chr>  <dbl>
#>  1   2020    15 00-0034… 83849bc5-0b6… 5161       Ezekiel Tu… LB    ARI      0  
#>  2   2020     1 00-0035… 23461354-f2d… 6778       Chris Stre… QB    ARI      0.3
#>  3   2020    17 00-0035… 23461354-f2d… 6778       Chris Stre… QB    ARI      7.4
#>  4   2020     1 00-0035… dd5a6b6e-ffa… 5849       Kyler Murr… QB    ARI     26.3
#>  5   2020     2 00-0035… dd5a6b6e-ffa… 5849       Kyler Murr… QB    ARI     32.1
#>  6   2020     3 00-0035… dd5a6b6e-ffa… 5849       Kyler Murr… QB    ARI     21.7
#>  7   2020     4 00-0035… dd5a6b6e-ffa… 5849       Kyler Murr… QB    ARI     23.1
#>  8   2020     5 00-0035… dd5a6b6e-ffa… 5849       Kyler Murr… QB    ARI     26.3
#>  9   2020     6 00-0035… dd5a6b6e-ffa… 5849       Kyler Murr… QB    ARI     28.9
#> 10   2020     7 00-0035… dd5a6b6e-ffa… 5849       Kyler Murr… QB    ARI     37.1
#> # ℹ 5,407 more rows
#> # ℹ 14 more variables: receiving_fumbles_lost <dbl>,
#> #   rushing_fumbles_lost <dbl>, sack_fumbles_lost <dbl>, interceptions <dbl>,
#> #   passing_2pt_conversions <dbl>, passing_tds <dbl>, passing_yards <dbl>,
#> #   receiving_2pt_conversions <dbl>, receiving_tds <dbl>,
#> #   receiving_yards <dbl>, receptions <dbl>, rushing_2pt_conversions <dbl>,
#> #   rushing_tds <dbl>, rushing_yards <dbl>
# }

# \donttest{
try({ # try only shown here because sometimes CRAN checks are weird
  template_conn <- ff_template(scoring_type = "sfb11", roster_type = "sfb11")
  ff_scoringhistory(template_conn, season = 2020)
}) # end try
#> # A tibble: 5,392 × 26
#>    season  week gsis_id    sportradar_id   mfl_id player_name pos   team  points
#>     <int> <int> <chr>      <chr>           <chr>  <chr>       <chr> <chr>  <dbl>
#>  1   2020     1 00-0035752 23461354-f2d7-… 15060  Chris Stre… QB    ARI      0.8
#>  2   2020    17 00-0035752 23461354-f2d7-… 15060  Chris Stre… QB    ARI     11.4
#>  3   2020     1 00-0035228 dd5a6b6e-ffae-… 14056  Kyler Murr… QB    ARI     39.3
#>  4   2020     2 00-0035228 dd5a6b6e-ffae-… 14056  Kyler Murr… QB    ARI     43.6
#>  5   2020     3 00-0035228 dd5a6b6e-ffae-… 14056  Kyler Murr… QB    ARI     31.7
#>  6   2020     4 00-0035228 dd5a6b6e-ffae-… 14056  Kyler Murr… QB    ARI     43.6
#>  7   2020     5 00-0035228 dd5a6b6e-ffae-… 14056  Kyler Murr… QB    ARI     40.8
#>  8   2020     6 00-0035228 dd5a6b6e-ffae-… 14056  Kyler Murr… QB    ARI     39.4
#>  9   2020     7 00-0035228 dd5a6b6e-ffae-… 14056  Kyler Murr… QB    ARI     60.6
#> 10   2020     9 00-0035228 dd5a6b6e-ffae-… 14056  Kyler Murr… QB    ARI     58.4
#> # ℹ 5,382 more rows
#> # ℹ 17 more variables: completions <dbl>, interceptions <dbl>,
#> #   passing_2pt_conversions <dbl>, passing_tds <dbl>, passing_yards <dbl>,
#> #   receiving_2pt_conversions <dbl>, receiving_tds <dbl>,
#> #   receiving_yards <dbl>, receptions <dbl>, rushing_2pt_conversions <dbl>,
#> #   rushing_first_downs <dbl>, rushing_tds <dbl>, rushing_yards <dbl>,
#> #   sacks <dbl>, special_teams_tds <dbl>, receiving_first_downs <dbl>, …
# }