ESPN Potential Points

This function calculates the optimal starters for a given week, using some lineup heuristics.

espn_potentialpoints(conn, weeks = 1:17)

Arguments

conn
the list object created by ff_connect()

weeks
a numeric vector for determining which weeks to calculate

Value

a tibble with the best lineup for each team and whether they were started or not

Examples

# \donttest{
try({ # try only shown here because sometimes CRAN checks are weird
  conn <- espn_connect(season = 2020, league_id = 899513)
  espn_potentialpoints(conn, weeks = 1:3)
}) # end try

#> # A tibble: 666 × 11
#>     week franchise_id franchise_name franchise_score optimal_slot actual_slot
#>    <int>        <int> <chr>                    <dbl> <chr>        <chr>      
#>  1     1            1 The Early GGod            102. QB           OP         
#>  2     1            1 The Early GGod            102. RB           BE         
#>  3     1            1 The Early GGod            102. RB           RB         
#>  4     1            1 The Early GGod            102. WR           BE         
#>  5     1            1 The Early GGod            102. WR           BE         
#>  6     1            1 The Early GGod            102. TE           BE         
#>  7     1            1 The Early GGod            102. OP           WR         
#>  8     1            1 The Early GGod            102. DST          DST        
#>  9     1            1 The Early GGod            102. K            K          
#> 10     1            1 The Early GGod            102. RB/WR/TE     BE         
#> # … with 656 more rows, and 5 more variables: player_score <dbl>,
#> #   player_name <chr>, player_pos <chr>, team <chr>, player_id <int>

# }