Get Draft Picks

Returns all draft picks (current and future) that belong to a specific franchise and have not yet been converted into players (i.e. selected.)

ff_draftpicks(conn, ...)

# S3 method for flea_conn
ff_draftpicks(conn, franchise_id = NULL, ...)

# S3 method for mfl_conn
ff_draftpicks(conn, ...)

# S3 method for sleeper_conn
ff_draftpicks(conn, ...)

Arguments

conn
the list object created by ff_connect()

...
other arguments (currently unused)

franchise_id
A list of franchise IDs to pull, if NULL will return all franchise IDs

Value

Returns a dataframe with current and future draft picks for each franchise

Methods (by class)

flea_conn: Fleaflicker: retrieves current and future draft picks, potentially for a specified team.

mfl_conn: MFL: returns current and future picks

sleeper_conn: Sleeper: retrieves current and future draft picks

Examples

# \donttest{
conn <- fleaflicker_connect(2020,206154)
ff_draftpicks(conn, franchise_id = 1373475)
# }

# \donttest{
dlf_conn <- mfl_connect(2020, league_id = 37920)
ff_draftpicks(conn = dlf_conn)

#> # A tibble: 80 x 5
#>    season franchise_id franchise_name round original_franchise_id
#>     <dbl> <chr>        <chr>          <dbl> <chr>                
#>  1   2021 0001         @dibari22          1 0001                 
#>  2   2021 0001         @dibari22          2 0001                 
#>  3   2021 0001         @dibari22          3 0001                 
#>  4   2021 0001         @dibari22          5 0001                 
#>  5   2021 0001         @dibari22          2 0009                 
#>  6   2021 0002         @_PeteLaw          1 0002                 
#>  7   2021 0002         @_PeteLaw          2 0002                 
#>  8   2021 0002         @_PeteLaw          3 0002                 
#>  9   2021 0002         @_PeteLaw          4 0002                 
#> 10   2021 0002         @_PeteLaw          5 0002                 
#> # … with 70 more rows

# }

# \donttest{
jml_conn <- ff_connect(platform = "sleeper", league_id = 522458773317046272, season = 2020)
ff_draftpicks(jml_conn)

#> # A tibble: 144 x 5
#>    season franchise_id franchise_name round original_franchise_id
#>    <chr>         <int> <chr>          <int>                 <int>
#>  1 2021              1 Fake News          3                     1
#>  2 2021              1 Fake News          4                     1
#>  3 2022              1 Fake News          2                     1
#>  4 2022              1 Fake News          3                     1
#>  5 2022              1 Fake News          4                     1
#>  6 2023              1 Fake News          1                     1
#>  7 2023              1 Fake News          2                     1
#>  8 2023              1 Fake News          3                     1
#>  9 2023              1 Fake News          4                     1
#> 10 2021              2 KingGabe           1                     2
#> # … with 134 more rows

# }