Get League Summary

This function returns a tidy dataframe of common league settings, including details like "1QB" or "2QB/SF", scoring, best ball, team count, IDP etc. This is potentially useful in summarising the features of multiple leagues.

ff_league(conn)

# S3 method for espn_conn
ff_league(conn)

# S3 method for flea_conn
ff_league(conn)

# S3 method for mfl_conn
ff_league(conn)

# S3 method for sleeper_conn
ff_league(conn)

Arguments

conn
the connection object created by ff_connect()

Value

A one-row summary of each league's main features.

Methods (by class)

espn_conn: ESPN: returns a summary of league features.

flea_conn: Flea: returns a summary of league features.

mfl_conn: MFL: returns a summary of league features.

sleeper_conn: Sleeper: returns a summary of league features.

Examples

# \donttest{
conn <- espn_connect(season = 2020, league_id = 899513)
ff_league(conn)

#> # A tibble: 1 x 15
#>   league_id league_name league_type franchise_count qb_type idp   scoring_flags
#>   <chr>     <chr>       <chr>                 <int> <chr>   <lgl> <chr>        
#> 1 899513    Sucio Boys  keeper                   10 2QB/SF  FALSE 0.5_ppr      
#> # … with 8 more variables: best_ball <lgl>, salary_cap <lgl>,
#> #   player_copies <dbl>, years_active <chr>, qb_count <chr>, roster_size <int>,
#> #   league_depth <dbl>, keeper_count <int>

# }

# \donttest{
conn <- fleaflicker_connect(2020, 206154)
ff_league(conn)

#> # A tibble: 1 x 14
#>   league_id league_name  league_type franchise_count qb_type idp   scoring_flags
#>   <chr>     <chr>        <chr>                 <dbl> <chr>   <lgl> <chr>        
#> 1 206154    Jackpot Dyn… dynasty                  16 1QB     TRUE  0.5_ppr, TEP…
#> # … with 7 more variables: best_ball <lgl>, salary_cap <lgl>,
#> #   player_copies <dbl>, qb_count <chr>, roster_size <int>, league_depth <dbl>,
#> #   keeper_count <int>

# }

# \donttest{
ssb_conn <- ff_connect(platform = "mfl", league_id = 54040, season = 2020)
ff_league(ssb_conn)

#> # A tibble: 1 x 13
#>   league_id league_name    franchise_count qb_type idp   scoring_flags best_ball
#>   <chr>     <chr>                    <dbl> <chr>   <lgl> <chr>         <lgl>    
#> 1 54040     The Super Sma…              14 1QB     FALSE 0.5_ppr, TEP… FALSE    
#> # … with 6 more variables: salary_cap <lgl>, player_copies <dbl>,
#> #   years_active <chr>, qb_count <chr>, roster_size <dbl>, league_depth <dbl>

# }

# \donttest{
jml_conn <- ff_connect(platform = "sleeper", league_id = "522458773317046272", season = 2020)
ff_league(jml_conn)

#> # A tibble: 1 x 15
#>   league_id  league_name league_type franchise_count qb_type idp   scoring_flags
#>   <chr>      <chr>       <chr>                 <dbl> <chr>   <lgl> <chr>        
#> 1 522458773… The JanMic… dynasty                  12 1QB     FALSE 0.5_ppr      
#> # … with 8 more variables: best_ball <lgl>, salary_cap <lgl>,
#> #   player_copies <dbl>, years_active <chr>, qb_count <chr>, roster_size <int>,
#> #   league_depth <dbl>, prev_league_ids <chr>

# }