Get Standings

This function returns a tidy dataframe of season-long fantasy team stats, including H2H wins as well as points, potential points, and all-play.

ff_standings(conn, ...)

# S3 method for flea_conn
ff_standings(conn, include_allplay = TRUE, include_potentialpoints = TRUE, ...)

# S3 method for mfl_conn
ff_standings(conn, ...)

# S3 method for sleeper_conn
ff_standings(conn, ...)

Arguments

conn
a conn object created by ff_connect()

...
arguments passed to other methods (currently none)

include_allplay
TRUE/FALSE - return all-play win pct calculation? defaults to TRUE

include_potentialpoints
TRUE/FALSE - return potential points calculation? defaults to TRUE.

Value

A tidy dataframe of standings data

Methods (by class)

flea_conn: Fleaflicker: returns H2H/points/all-play/best-ball data in a table.

mfl_conn: MFL: returns H2H/points/all-play/best-ball data in a table.

sleeper_conn: Sleeper: returns all standings and points data and manually calculates allplay results.

Examples

# \donttest{
conn <- fleaflicker_connect(season = 2020, league_id = 206154)
x <- ff_standings(conn)
# }

# \donttest{
ssb_conn <- ff_connect(platform = "mfl", league_id = 54040, season = 2020)
ff_standings(ssb_conn)

#> # A tibble: 14 x 20
#>    franchise_id franchise_name h2h_wins h2h_losses h2h_ties h2h_winpct
#>    <chr>        <chr>             <dbl>      <dbl>    <dbl>      <dbl>
#>  1 0014         Team Luigi           10          3        0      0.769
#>  2 0009         Team Link            10          3        0      0.769
#>  3 0010         Team Yoshi            9          4        0      0.692
#>  4 0006         Team King Ded…        8          5        0      0.615
#>  5 0011         Team Diddy Ko…        8          5        0      0.615
#>  6 0007         Team Kirby            7          6        0      0.538
#>  7 0008         Team Bowser           7          6        0      0.538
#>  8 0003         Team Donkey K…        6          7        0      0.462
#>  9 0002         Team Simon Be…        6          7        0      0.462
#> 10 0004         Team Ice Clim…        5          8        0      0.385
#> 11 0005         Team Dr. Mario        5          8        0      0.385
#> 12 0012         Team Mewtwo           4          9        0      0.308
#> 13 0013         Team Ness             3         10        0      0.231
#> 14 0001         Team Pikachu          3         10        0      0.231
#> # … with 14 more variables: allplay_wins <dbl>, allplay_losses <dbl>,
#> #   allplay_ties <dbl>, allplay_winpct <dbl>, points_for <dbl>,
#> #   points_against <dbl>, max_points_against <dbl>, min_points_against <dbl>,
#> #   potential_points <dbl>, victory_points <dbl>, offensive_points <dbl>,
#> #   defensive_points <dbl>, power_rank <dbl>, power_rank_alt <dbl>

# }

# \donttest{
jml_conn <- ff_connect(platform = "sleeper", league_id = "522458773317046272", season = 2020)
ff_standings(jml_conn)

#> # A tibble: 12 x 12
#>    franchise_id franchise_name h2h_wins h2h_losses h2h_ties h2h_winpct
#>           <int> <chr>             <dbl>      <dbl>    <dbl>      <dbl>
#>  1            1 Fake News             8          5        0     0.615 
#>  2            2 KingGabe              1         12        0     0.0769
#>  3            3 solarpool             8          5        0     0.615 
#>  4            4 The FANTom Me…        8          5        0     0.615 
#>  5            5 Barbarians            6          7        0     0.462 
#>  6            6 sox05syd              8          5        0     0.615 
#>  7            7 Flipadelphia05       10          3        0     0.769 
#>  8            8 Hocka Flocka          7          6        0     0.538 
#>  9            9 ZPMiller97            4          9        0     0.308 
#> 10           10 JMLarkin              1         12        0     0.0769
#> 11           11 Permian Panth…        8          5        0     0.615 
#> 12           12 jaydk                 9          4        0     0.692 
#> # … with 6 more variables: points_for <dbl>, points_against <dbl>,
#> #   potential_points <dbl>, allplay_wins <dbl>, allplay_losses <dbl>,
#> #   allplay_winpct <dbl>

# }